+++ backend/Infrastructure/Migrations/AppDBContextModelSnapshot.cs
@@ -40,8 +40,8 @@ namespace Infrastructure.Migrations
                         .HasMaxLength(400)
                         .HasColumnType("nvarchar(400)");
 
-                    b.Property<int>("Status")
-                        .HasColumnType("int");
+                    b.Property<byte>("Status")
+                        .HasColumnType("tinyint");
 
                     b.Property<string>("Titulo")
                         .IsRequired()
